// client/src/setup/store/actionsTypes/test.actionsTypes.ts
export enum TestActionType {
    REQUEST_GET_TESTS = "REQUEST_GET_TESTS ",
    SUCCESS_GET_TESTS = "SUCCESS_GET_TESTS",
    FAILED_GET_TESTS = "FAILED_GET_TESTS",
    REQUEST_CREATE_TEST = "REQUEST_CREATE_TEST ",
    SUCCESS_CREATE_TEST = "SUCCESS_CREATE_TEST",
    FAILED_CREATE_TEST = "FAILED_CREATE_TEST",
    REQUEST_UPDATE_TEST = "REQUEST_UPDATE_TEST ",
    SUCCESS_UPDATE_TEST = "SUCCESS_UPDATE_TEST",
    FAILED_UPDATE_TEST = "FAILED_UPDATE_TEST",
    REQUEST_DELETE_TEST = "REQUEST_DELETE_TEST ",
    SUCCESS_DELETE_TEST = "SUCCESS_DELETE_TEST",
    FAILED_DELETE_TEST = "FAILED_DELETE_TEST",
  }
  